# agri_uav_disease_system/backend/docs/integration_package/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "三下乡无人机水稻病虫害识别系统",
    "version": "mvp-1"
  },
  "paths": {
...
    "/api/dashboard/plots": {
      "get": {
        "tags": [
          "dashboard"
        ],
        "summary": "Dashboard Plots",
        "operationId": "dashboard_plots_api_dashboard_plots_get",
        "parameters": [
          {
            "name": "region_name",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Region Name"
            }
          },
          {
            "name": "risk_level",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Risk Level"
            }
          },
          {
            "name": "disease",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Disease"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlotStatisticsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/dashboard/plots/{plot_id}": {
      "get": {
        "tags": [
          "dashboard"
        ],
        "summary": "Dashboard Plot Detail",
        "operationId": "dashboard_plot_detail_api_dashboard_plots__plot_id__get",
        "parameters": [
          {
            "name": "plot_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Plot Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlotDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/dashboard/plots/{plot_id}/records": {
      "get": {
        "tags": [
          "dashboard"
        ],
        "summary": "Dashboard Plot Records",
        "operationId": "dashboard_plot_records_api_dashboard_plots__plot_id__records_get",
        "parameters": [
          {
            "name": "plot_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Plot Id"
            }
          },
          {
            "name": "risk_level",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Risk Level"
            }
          },
          {
            "name": "severity",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Severity"
            }
          },
          {
            "name": "disease",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Disease"
            }
          },
          {
            "name": "start_time",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Start Time"
            }
          },
          {
            "name": "end_time",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "End Time"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecordListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/dashboard/heatmap": {
      "get": {
        "tags": [
          "dashboard"
        ],
        "summary": "Dashboard Heatmap",
        "operationId": "dashboard_heatmap_api_dashboard_heatmap_get",
        "parameters": [
          {
            "name": "region_name",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Region Name"
            }
          },
          {
            "name": "disease",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Disease"
            }
          },
          {
            "name": "risk_level",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Risk Level"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HeatmapResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/dashboard/latest-records": {
      "get": {
        "tags": [
          "dashboard"
        ],
        "summary": "Dashboard Latest Records",
        "operationId": "dashboard_latest_records_api_dashboard_latest_records_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "default": 10,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LatestRecordsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/dashboard/latest-alerts": {
      "get": {
        "tags": [
          "dashboard"
        ],
        "summary": "Dashboard Latest Alerts",
        "operationId": "dashboard_latest_alerts_api_dashboard_latest_alerts_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "default": 10,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LatestAlertsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/agent/knowledge-context": {
      "post": {
        "tags": [
          "experimental-agent"
        ],
        "summary": "Knowledge Context",
        "operationId": "knowledge_context_api_agent_knowledge_context_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KnowledgeContextRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeContextResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/agent/diagnosis-report": {
      "post": {
        "tags": [
          "experimental-agent"
        ],
        "summary": "Diagnosis Report",
        "operationId": "diagnosis_report_api_agent_diagnosis_report_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DiagnosisReportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiagnosisReportResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DiagnosisReportRequest": {
        "properties": {
          "record_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Record Id"
          },
          "disease_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 80
              },
              {
                "type": "null"
              }
            ],
            "title": "Disease Id"
          },
          "model_class": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 120
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Class"
          },
          "confidence": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 1.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence"
          },
          "source_type": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 80
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Type"
          },
          "user_question": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "User Question"
          },
          "field_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 120
              },
              {
                "type": "null"
              }
            ],
            "title": "Field Id"
          },
          "plot_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 120
              },
              {
                "type": "null"
              }
            ],
            "title": "Plot Id"
          },
          "uav_task_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 120
              },
              {
                "type": "null"
              }
            ],
            "title": "Uav Task Id"
          },
          "abnormal_region_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 120
              },
              {
                "type": "null"
              }
            ],
            "title": "Abnormal Region Id"
          },
          "risk_level": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 40
              },
              {
                "type": "null"
              }
            ],
            "title": "Risk Level"
          },
          "severity": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 80
              },
              {
                "type": "null"
              }
            ],
            "title": "Severity"
          }
        },
        "type": "object",
        "title": "DiagnosisReportRequest"
      },
...
      "Geo": {
        "properties": {
          "lng": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lng"
          },
          "lat": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lat"
          }
        },
        "type": "object",
        "title": "Geo"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "HeatmapResponse": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type",
            "default": "heatmap_data"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "points": {
            "items": {
              "$ref": "#/components/schemas/HeatmapPoint"
            },
            "type": "array",
            "title": "Points"
          }
        },
        "type": "object",
        "required": [
          "total",
          "points"
        ],
        "title": "HeatmapResponse"
      },
...
      "KnowledgeContextRequest": {
        "properties": {
          "question": {
            "type": "string",
            "maxLength": 500,
            "minLength": 1,
            "title": "Question"
          },
          "record_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 120
              },
              {
                "type": "null"
              }
            ],
            "title": "Record Id"
          },
          "disease_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 80
              },
              {
                "type": "null"
              }
            ],
            "title": "Disease Id"
          },
          "model_class": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 120
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Class"
          },
          "confidence": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 1.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence"
          },
          "source_type": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 80
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Type"
          },
          "top_k": {
            "type": "integer",
            "maximum": 20.0,
            "minimum": 1.0,
            "title": "Top K",
            "default": 5
          },
          "include_knowledge_chunks": {
            "type": "boolean",
            "title": "Include Knowledge Chunks",
            "default": true
          },
          "include_graph": {
            "type": "boolean",
            "title": "Include Graph",
            "default": true
          },
          "include_relations": {
            "type": "boolean",
            "title": "Include Relations",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "question"
        ],
        "title": "KnowledgeContextRequest"
      },
      "KnowledgeContextResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "mode": {
            "type": "string",
            "title": "Mode",
            "default": "knowledge_context"
          },
          "question": {
            "type": "string",
            "title": "Question"
          },
          "matched_disease": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/KnowledgeContextDisease"
              },
              {
                "type": "null"
              }
            ]
          },
          "knowledge_chunks": {
            "items": {
              "$ref": "#/components/schemas/KnowledgeContextChunk"
            },
            "type": "array",
            "title": "Knowledge Chunks"
          },
          "graph": {
            "$ref": "#/components/schemas/KnowledgeContextGraph"
          },
          "context_summary": {
            "type": "string",
            "title": "Context Summary"
          },
          "safety_notice": {
            "type": "string",
            "title": "Safety Notice"
          },
          "insufficient_evidence": {
            "type": "boolean",
            "title": "Insufficient Evidence",
            "default": false
          },
          "missing_context": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Missing Context"
          }
        },
        "type": "object",
        "required": [
          "question",
          "context_summary",
          "safety_notice"
        ],
        "title": "KnowledgeContextResponse"
      },
...
      "LatestAlertsResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/LatestAlertItem"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "LatestAlertsResponse"
      },
...
      "LatestRecordsResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/LatestRecordItem"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "LatestRecordsResponse"
      },
...
      "PlotDetailResponse": {
        "properties": {
          "plot_id": {
            "type": "string",
            "title": "Plot Id"
          },
          "plot_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Plot Name"
          },
          "region_name": {
            "type": "string",
            "title": "Region Name"
          },
          "geo": {
            "$ref": "#/components/schemas/Geo"
          },
          "record_count": {
            "type": "integer",
            "title": "Record Count"
          },
          "disease_record_count": {
            "type": "integer",
            "title": "Disease Record Count"
          },
          "normal_record_count": {
            "type": "integer",
            "title": "Normal Record Count"
          },
          "main_disease": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Main Disease"
          },
          "disease_types": {
            "items": {
              "$ref": "#/components/schemas/DiseaseTypeCount"
            },
            "type": "array",
            "title": "Disease Types"
          },
          "max_severity": {
            "type": "string",
            "title": "Max Severity"
          },
          "risk_level": {
            "type": "string",
            "title": "Risk Level"
          },
          "latest_detect_time": {
            "type": "string",
            "title": "Latest Detect Time"
          },
          "latest_record": {
            "$ref": "#/components/schemas/DetectionResult"
          },
          "latest_alert": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Latest Alert"
          },
          "suggestion_summary": {
            "type": "string",
            "title": "Suggestion Summary"
          }
        },
        "type": "object",
        "required": [
          "plot_id",
          "region_name",
          "geo",
          "record_count",
          "disease_record_count",
          "normal_record_count",
          "disease_types",
          "max_severity",
          "risk_level",
          "latest_detect_time",
          "latest_record",
          "suggestion_summary"
        ],
        "title": "PlotDetailResponse"
      },
      "PlotStatisticItem": {
        "properties": {
          "plot_id": {
            "type": "string",
            "title": "Plot Id"
          },
          "plot_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Plot Name"
          },
          "region_name": {
            "type": "string",
            "title": "Region Name"
          },
          "record_count": {
            "type": "integer",
            "title": "Record Count"
          },
          "disease_record_count": {
            "type": "integer",
            "title": "Disease Record Count"
          },
          "main_disease": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Main Disease"
          },
          "max_severity": {
            "type": "string",
            "title": "Max Severity"
          },
          "risk_level": {
            "type": "string",
            "title": "Risk Level"
          },
          "latest_detect_time": {
            "type": "string",
            "title": "Latest Detect Time"
          },
          "geo": {
            "$ref": "#/components/schemas/Geo"
          }
        },
        "type": "object",
        "required": [
          "plot_id",
          "region_name",
          "record_count",
          "disease_record_count",
          "max_severity",
          "risk_level",
          "latest_detect_time",
          "geo"
        ],
        "title": "PlotStatisticItem"
      },
      "PlotStatisticsResponse": {
        "properties": {
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/PlotStatisticItem"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "total",
          "items"
        ],
        "title": "PlotStatisticsResponse"
      },
...
      "RecordListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/DetectionResult"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "page_size"
        ],
        "title": "RecordListResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}